+++ packages/client/src/lib/drizzle/schema.ts
@@ -1,4 +1,4 @@
-export * from "./schema/codeUpdates";
+export * from "./schema/directorUpdates";
 export * from "./schema/llamaInferences";
 export * from "./schema/simulations";
-export * from "./schema/storyUpdates";
+export * from "./schema/writerUpdates";