+++ packages/client/src/lib/drizzle/schema/simulations.ts
@@ -2,7 +2,7 @@ import { sortByKey } from "@/lib/utils";
 import { relations, sql } from "drizzle-orm";
 import { sqliteTable, text } from "drizzle-orm/sqlite-core";
 import { nanoid } from "nanoid";
-import { storyUpdates } from "./storyUpdates";
+import { writerUpdates } from "./writerUpdates";
 
 export const simulations = sqliteTable(
   "simulations",
@@ -12,7 +12,7 @@ export const simulations = sqliteTable(
       .$defaultFn(() => nanoid()),
 
     scenarioId: text("scenario_id").notNull(),
-    headStoryUpdateId: text("head_story_update_id"),
+    headWriterUpdateId: text("head_writer_update_id"),
 
     createdAt: text("created_at")
       .notNull()
@@ -25,9 +25,9 @@ export const simulations = sqliteTable(
 );
 
 export const simulationRelatiosn = relations(simulations, ({ one, many }) => ({
-  storyUpdates: many(storyUpdates),
-  headStoryUpdate: one(storyUpdates, {
-    fields: [simulations.headStoryUpdateId],
-    references: [storyUpdates.id],
+  writerUpdates: many(writerUpdates),
+  headWriterUpdate: one(writerUpdates, {
+    fields: [simulations.headWriterUpdateId],
+    references: [writerUpdates.id],
   }),
 }));